[PI-1] 두 화면에서 공통으로 사용하는 TitleView 분리
