+++ Planz/Management/Standby/StandbyCellView.swift
@@ -14,23 +14,25 @@ struct StandbyCellView: View {
     
     var body: some View {
         Group {
-            VStack {
-                HStack {
-                    Text(item.title)
-                        .font(.title3)
-                        .bold()
-                        .border(.orange)
-                    RoleMarkView(role: item.role)
+            HStack {
+                VStack {
+                    ManagementTitleCellView(
+                        title: item.title,
+                        role: item.role
+                    )
+                    
+                    HStack {
+                        Text(item.leaderName)
+                        Rectangle()
+                            .frame(width: 1, height: 12)
+                        Text("\(item.replyPeopleCount)명 응답완료")
+                    }
+                    .foregroundColor(.gray)
+                    .frame(maxWidth: .infinity, alignment: .leading)
                 }
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .border(.black)
-                
-                HStack {
-                    Text(item.leaderName)
-                    Text("\(item.replyPeopleCount)명 응답완료")
-                }
-                .foregroundColor(.gray)
-                .frame(maxWidth: .infinity, alignment: .leading)
+                Image(systemName: "chevron.forward")
+                    .frame(width: 32, height: 32)
+                    .foregroundColor(.gray)
             }
             .border(.gray)
             .padding()
@@ -46,7 +48,13 @@ struct StandbyCellView: View {
 struct StandbyCellView_Previews: PreviewProvider {
     static var previews: some View {
         StandbyCellView(item:
-                .constant(.init(title: "Title-1", role: .general, leaderName: "LeaderName", replyPeopleCount: 5))
+                .constant(
+                    .init(
+                    title: "가나다라마바사아자차카파타하이",
+                    role: .general,
+                    leaderName: "LeaderName",
+                    replyPeopleCount: 5)
+                )
         )
     }
 }
